# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SplitApp Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/home/summary": {
      "get": {
        "tags": [
          "Home"
        ],
        "summary": "Get Home Summary",
        "operationId": "get_home_summary_api_home_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HomeMoneyBucket": {
        "properties": {
          "owed_kopecks": {
            "type": "integer",
            "title": "Owed Kopecks"
          },
          "receivable_kopecks": {
            "type": "integer",
            "title": "Receivable Kopecks"
          }
        },
        "type": "object",
        "required": [
          "owed_kopecks",
          "receivable_kopecks"
        ],
        "title": "HomeMoneyBucket"
      },
      "HomeSummary": {
        "properties": {
          "confirmed": {
            "$ref": "#/components/schemas/HomeMoneyBucket"
          },
          "pending": {
            "$ref": "#/components/schemas/HomeMoneyBucket"
          },
          "disputed": {
            "$ref": "#/components/schemas/HomeMoneyBucket"
          }
        },
        "type": "object",
        "required": [
          "confirmed",
          "pending",
          "disputed"
        ],
        "title": "HomeSummary"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}